+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Add combinedId to OHIFSeriesMEtadata to account for subseries - [@Triet]
+### Changed
+- Changed updateSeries and createDisplaySets to account for combinedId - [@Triet]
+
 ## [1.1.0] - 2020-08-06
 ### Fixed
 - Fix splitSeries does not return origin series when first load cause no image instance - [@tunglt1810]